+++ src/api/main.ts
@@ -6,7 +6,7 @@ export const scheduleList = async () => {
   return data;
 };
 
-export const noticeList = async () => {
+export const mainNoticeList = async () => {
   const { data } = await axios.get(`${MOCK_API_BASE_URL}/main/notice/list`);
   console.log(data);
   return data;
