+++ src/components/main/Notice/Notice.tsx
@@ -5,11 +5,11 @@ import NoticeItem from "./NoticeItem";
 import styled from "styled-components";
 import { useRouter } from "next/navigation";
 import Link from "@/components/common/Link/link";
-import { noticeListQuery } from "@/features/main";
+import { mainNoticeListQuery } from "@/features/main";
 
 const Notice = () => {
   const router = useRouter();
-  const { data } = noticeListQuery();
+  const { data } = mainNoticeListQuery();
 
   return (
     <StyledNotice>
