+++ src/features/main.ts
@@ -1,4 +1,4 @@
-import { noticeList, scheduleList } from "@/api/main";
+import { mainNoticeList, scheduleList } from "@/api/main";
 import { useQuery } from "react-query";
 
 /** 메인 입학 일정 리스트 */
@@ -17,14 +17,14 @@ export const scheduleListQuery = () => {
 
 /** 메인 공지사항 리스트 */
 
-interface NoticeListType {
+interface MainNoticeListType {
   id: number;
   title: string;
 }
 
-export const noticeListQuery = () => {
-  const { data } = useQuery<NoticeListType[]>(["noticeList"], () =>
-    noticeList()
+export const mainNoticeListQuery = () => {
+  const { data } = useQuery<MainNoticeListType[]>(["mainNoticeList"], () =>
+    mainNoticeList()
   );
   return { data: data || [] };
 };
